visualization/requirements.txt
flask
flask-cors
pillow
opencv-python
